requirements.txt
Django==1.11.5
django-filter==1.0.4
djangorestframework==3.6.4
Markdown==2.6.9
pytz==2017.2
drf-writable-nested==0.3.1
mysql-connector==2.1.4
mysqlclient==1.3.10